Added new slash command for SpaceX launches
Also added a function to get upcoming SpaceX launches from an API. The API is restricted to only 15 queries per hour so I am currently using the dev API they provide which is unlimited but has old data. I will simply switch the API URL to the correct one for full deployment.

Because of the very limited API requests they provide I have build a function that simply queries it once every 2 hours and then stores the JSON data locally, this data is then used whenever a user makes a request for the SpaceX launch data.